+++ clearstats-api/src/database/statistics/upsert_vote.rs
@@ -1,11 +1,11 @@
-use shared_lib::database::{DatabaseInteger, DatabaseResult};
 use shared_lib::database::manager::{DatabaseManager, DatabaseManagerTrait};
+use shared_lib::database::{DatabaseInteger, DatabaseResult};
 
 pub async fn run_query(
     db_manager: &DatabaseManager,
     statistic_id: DatabaseInteger,
     user_id: DatabaseInteger,
-    vote: i8,
+    vote: i16,
 ) -> DatabaseResult {
     let pool = db_manager.get_database_pool();
